requirements.txt
# Core
Discord.py>=2.3.2,<3.0
aiosqlite>=0.20.0,<1.0
python-dotenv>=1.0.0,<2.0
aiohttp>=3.9.0,<4.0
PyNaCl>=1.5.0,<2.0

# AI
openai>=1.40.0,<2.0
